refactor: restructure `gathermaputil` package to `gathermap` and update references
